requirements.txt
async-timeout==4.0.3
click==8.1.7
hiredis==3.0.0
joblib==1.4.2
nltk==3.9.1
numpy==2.0.1
psutil==6.0.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
regex==2024.7.24
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sparse-dot-topn==1.1.3
threadpoolctl==3.5.0
tqdm==4.66.5
docker==7.1.0
